Fixing DataTypes and removing Tempid

+++ DevTrivia/DevTrivia.API/Capabilities/PlayerStats/Database/Entities/PlayerStats.cs
@@ -7,8 +7,8 @@ namespace DevTrivia.API.Capabilities.PlayerStats.Database.Entities;
 public sealed class PlayerStatsEntity : BaseEntity
 {
     public required long UserId { get; set; }
-    public long TotalMatches { get; set; } = 1;
-    public long TotalCorrect { get; set; }
+    public int TotalMatches { get; set; } = 1;
+    public int TotalCorrect { get; set; }
     public EloRating EloRating { get; set; }
 
     public UserEntity User { get; set; } = null!;
