+++ DevTrivia/DevTrivia.API/Capabilities/PlayerStats/Models/PlayerStatsRequest.cs
@@ -5,6 +5,6 @@ namespace DevTrivia.API.Capabilities.PlayerStats.Models;
 public record PlayerStatsRequest
 {
     public required long UserId { get; init; }
-    public required long TotalCorrect { get; init; }
+    public required int TotalCorrect { get; init; }
     public required EloRating EloRating { get; init; }
 }
